+++ eccorce_app/app/src/main/java/com/example/eccomerce_app/viewModel/PaymentViewModel.kt
@@ -8,7 +8,7 @@ import com.example.eccomerce_app.dto.StripeClientSecret
 class PaymentViewModel(private val paymentRepository: PaymentRepository): ViewModel() {
 
 
-    suspend fun  submitOrderToStripe(totalPrice: Double):String?{
+    suspend fun  submitOrderToStripe(totalPrice: Long):String?{
       return  when(val result = paymentRepository.generatePaymentIntent(totalPrice)){
             is NetworkCallHandler.Successful<*> -> {
                 val data = result.data as StripeClientSecret
